+++ src/lib/hono/hono.client.ts
@@ -1,0 +1,4 @@
+import type { AppType } from "./hono.server";
+import { hc } from "hono/client";
+
+export const honoClient = hc<AppType>("http://localhost:5173/");
